+++ .gitignore
@@ -41,3 +41,7 @@ releases/
 
 # Windows logo .syso files
 *.syso
+
+# Fyne-cross build
+fyne-cross/
+Icon.png
