+++ app/manager/supervisor_mobile.go
@@ -4,15 +4,8 @@ package manager
 
 import (
 	"github.com/LEX0RE/rockpload/app/config"
-	"github.com/LEX0RE/rockpload/app/rocket_network"
 	"github.com/LEX0RE/rockpload/app/tools"
 	"github.com/LEX0RE/rockpload/app/tools/logger"
-)
-
-const (
-	EVENT_ON_RL_DETECTED        tools.EventType = "rl_detected"
-	EVENT_ON_RL_PLAYER_DETECTED tools.EventType = "rl_player_detected"
-	EVENT_ON_RL_CLOSED          tools.EventType = "rl_closed"
 )
 
 type RLSupervisor struct {
@@ -20,16 +13,6 @@ type RLSupervisor struct {
 
 	EventManager *tools.EventManager
 	RLLogInfo    *RLLogInfo
-}
-
-type RLVersionInfo struct {
-	GameVersion string
-	FeatureSet  string
-}
-
-type RLLogInfo struct {
-	VersionInfo  *RLVersionInfo
-	AccountFound *rocket_network.Account
 }
 
 func NewRLSupervisor(appConfig *config.AppConfig, accountManager *AccountManager) *RLSupervisor {
@@ -42,8 +25,21 @@ func NewRLSupervisor(appConfig *config.AppConfig, accountManager *AccountManager
 	}
 }
 
-func (rls *RLSupervisor) Start() {}
+func (rls *RLSupervisor) Supervise() {
+	logger.FuncDebug()
+}
 
-func (rls *RLSupervisor) Stop() {}
+// Looper Inherit method
+func (rls *RLSupervisor) Start() {
+	logger.FuncDebug()
+}
 
-func (rls *RLSupervisor) Toggle(bool) {}
+// Looper Inherit method
+func (rls *RLSupervisor) Stop() {
+	logger.FuncDebug()
+}
+
+// Looper Inherit method
+func (rls *RLSupervisor) Toggle(bool) {
+	logger.FuncDebug()
+}
