+++ app/tools/browser_mobile.go
@@ -62,5 +62,7 @@ func ClearBrowserProfile(profileId int) {
 }
 
 func getBrowserProfilePath(profileId int) string {
+	logger.FuncDebug()
+
 	return BrowserProfilePrefix + strconv.Itoa(profileId)
 }
